refactor: 디렉토리 정리 및 컴포넌트 위치 조정
